fix: вынести загрузку компонентов для lazy в отдельный файл
