+++ src/routes/index.tsx
@@ -1,20 +1,21 @@
-import { lazy } from 'react';
 import {
   Route,
   createBrowserRouter,
   createRoutesFromElements,
 } from 'react-router-dom';
 
+import {
+  AmbassadorsPage,
+  CandidatesPage,
+  LoginPage,
+  MerchPage,
+  NotFoundPage,
+  NotificationsPage,
+  StatisticsPage,
+} from './components';
+
 import Layout from '../components/Layout';
 import ProtectedRoute from '../components/ProtectedRoute';
-
-const AmbassadorsPage = lazy(() => import('../pages/AmbassadorsPage'));
-const CandidatesPage = lazy(() => import('../pages/CandidatesPage'));
-const MerchPage = lazy(() => import('../pages/MerchPage'));
-const NotFoundPage = lazy(() => import('../pages/NotFoundPage'));
-const NotificationsPage = lazy(() => import('../pages/NotificationsPage'));
-const StatisticsPage = lazy(() => import('../pages/StatisticsPage'));
-const LoginPage = lazy(() => import('../pages/LoginPage'));
 
 const router = createBrowserRouter(
   createRoutesFromElements(
